# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Score {
  opportunity: number;
  risk: number;
  confidence: number;
}

export type EvidenceType = typeof EvidenceType[keyof typeof EvidenceType];


export const EvidenceType = {
  FACT: 'FACT',
  INFERENCE: 'INFERENCE',
  UNKNOWN: 'UNKNOWN',
} as const;

export interface Evidence {
  type: EvidenceType;
  label: string;
  value: string;
  source: string;
}

export type RiskFlagSeverity = typeof RiskFlagSeverity[keyof typeof RiskFlagSeverity];


export const RiskFlagSeverity = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL',
} as const;

export interface RiskFlag {
  severity: RiskFlagSeverity;
  label: string;
  detail: string;
}

export type DiscoveryCreatorRisk = typeof DiscoveryCreatorRisk[keyof typeof DiscoveryCreatorRisk];


export const DiscoveryCreatorRisk = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL',
} as const;

export type DiscoveryStatus = typeof DiscoveryStatus[keyof typeof DiscoveryStatus];


export const DiscoveryStatus = {
  IGNORE: 'IGNORE',
  WATCH: 'WATCH',
  INVESTIGATE: 'INVESTIGATE',
  'PAPER-TRADE': 'PAPER-TRADE',
  REVIEW: 'REVIEW',
} as const;

export interface Discovery {
  id: string;
  symbol: string;
  name: string;
  mint: string;
  venue: string;
  discoveredAt: string;
  ageMinutes: number;
  price: number;
  marketCap: number;
  liquidity: number;
  volume5m: number;
  buys: number;
  sells: number;
  holders: number;
  top10Concentration: number;
  creatorRisk: DiscoveryCreatorRisk;
  momentum: number;
  status: DiscoveryStatus;
  scores: Score;
  evidence: Evidence[];
  risks: RiskFlag[];
  unknowns: string[];
  reasoning: string;
  nextAction: string;
}

export type AgentOverviewMode = typeof AgentOverviewMode[keyof typeof AgentOverviewMode];


export const AgentOverviewMode = {
  PAPER: 'PAPER',
  'LIVE-DATA': 'LIVE-DATA',
} as const;

export type AgentOverviewStatus = typeof AgentOverviewStatus[keyof typeof AgentOverviewStatus];


export const AgentOverviewStatus = {
  RUNNING: 'RUNNING',
  DEGRADED: 'DEGRADED',
  PAUSED: 'PAUSED',
} as const;

export type SourceHealthStatus = typeof SourceHealthStatus[keyof typeof SourceHealthStatus];


export const SourceHealthStatus = {
  HEALTHY: 'HEALTHY',
  DEGRADED: 'DEGRADED',
  OFFLINE: 'OFFLINE',
} as const;

export interface SourceHealth {
  name: string;
  status: SourceHealthStatus;
  latencyMs: number;
  lastChecked: string;
}

export interface Performance {
  equity: number;
  returnPct: number;
  winRate: number;
  maxDrawdown: number;
  paperTrades: number;
  falsePositiveRate: number;
}

export interface AgentOverview {
  mode: AgentOverviewMode;
  status: AgentOverviewStatus;
  loopStage: string;
  lastEventAt: string;
  latencyMs: number;
  discoveries: number;
  investigate: number;
  paperTrade: number;
  decisionsToday: number;
  sourceHealth: SourceHealth[];
  performance: Performance;
}

export type ReplayInputDataset = typeof ReplayInputDataset[keyof typeof ReplayInputDataset];


export const ReplayInputDataset = {
  DEMO_24H: 'DEMO_24H',
  DEMO_7D: 'DEMO_7D',
} as const;

export interface ReplayInput {
  dataset: ReplayInputDataset;
  /** @minimum 1 */
  startingCapital: number;
  /** @minimum 0 */
  slippageBps: number;
  /** @minimum 0 */
  feeBps: number;
  /** @minimum 0 */
  latencyMs: number;
}

export interface ReplayResult {
  runId: string;
  dataset: string;
  startedAt: string;
  completedAt: string;
  seed: number;
  trades: number;
  filled: number;
  partialFills: number;
  failures: number;
  endingEquity: number;
  returnPct: number;
  maxDrawdown: number;
  notes: string[];
}

export type DecisionLogEntryDecision = typeof DecisionLogEntryDecision[keyof typeof DecisionLogEntryDecision];


export const DecisionLogEntryDecision = {
  IGNORE: 'IGNORE',
  WATCH: 'WATCH',
  INVESTIGATE: 'INVESTIGATE',
  'PAPER-TRADE': 'PAPER-TRADE',
  REVIEW: 'REVIEW',
} as const;

export interface DecisionLogEntry {
  id: string;
  createdAt: string;
  discoveryId: string;
  decision: DecisionLogEntryDecision;
  scores: Score;
  reasoning: string;
  evidenceCount: number;
  riskCount: number;
  unknownCount: number;
  nextAction: string;
  outcome: string;
  predictionAccuracy: number;
  signalPerformance: number;
}

export type SimulationStatus = typeof SimulationStatus[keyof typeof SimulationStatus];


export const SimulationStatus = {
  COMPLETED: 'COMPLETED',
  RUNNING: 'RUNNING',
  FAILED: 'FAILED',
} as const;

export interface Simulation {
  id: string;
  label: string;
  dataset: string;
  status: SimulationStatus;
  returnPct: number;
  maxDrawdown: number;
  winRate: number;
  trades: number;
  capital: number;
  createdAt: string;
}

export type ListDiscoveriesParams = {
status?: ListDiscoveriesStatus;
search?: string;
};

export type ListDiscoveriesStatus = typeof ListDiscoveriesStatus[keyof typeof ListDiscoveriesStatus];


export const ListDiscoveriesStatus = {
  ALL: 'ALL',
  IGNORE: 'IGNORE',
  WATCH: 'WATCH',
  INVESTIGATE: 'INVESTIGATE',
  'PAPER-TRADE': 'PAPER-TRADE',
  REVIEW: 'REVIEW',
} as const;