# lib/api-zod/src/generated/types/agentOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentOverviewMode } from './agentOverviewMode';
import type { AgentOverviewStatus } from './agentOverviewStatus';
import type { Performance } from './performance';
import type { SourceHealth } from './sourceHealth';

export interface AgentOverview {
  mode: AgentOverviewMode;
  status: AgentOverviewStatus;
  loopStage: string;
  lastEventAt: string;
  latencyMs: number;
  discoveries: number;
  investigate: number;
  paperTrade: number;
  decisionsToday: number;
  sourceHealth: SourceHealth[];
  performance: Performance;
}