# lib/api-zod/src/generated/types/decisionLogEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionLogEntryDecision } from './decisionLogEntryDecision';
import type { Score } from './score';

export interface DecisionLogEntry {
  id: string;
  createdAt: string;
  discoveryId: string;
  decision: DecisionLogEntryDecision;
  scores: Score;
  reasoning: string;
  evidenceCount: number;
  riskCount: number;
  unknownCount: number;
  nextAction: string;
  outcome: string;
  predictionAccuracy: number;
  signalPerformance: number;
}