# lib/api-zod/src/generated/types/discovery.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiscoveryCreatorRisk } from './discoveryCreatorRisk';
import type { DiscoveryStatus } from './discoveryStatus';
import type { Evidence } from './evidence';
import type { RiskFlag } from './riskFlag';
import type { Score } from './score';

export interface Discovery {
  id: string;
  symbol: string;
  name: string;
  mint: string;
  venue: string;
  discoveredAt: string;
  ageMinutes: number;
  price: number;
  marketCap: number;
  liquidity: number;
  volume5m: number;
  buys: number;
  sells: number;
  holders: number;
  top10Concentration: number;
  creatorRisk: DiscoveryCreatorRisk;
  momentum: number;
  status: DiscoveryStatus;
  scores: Score;
  evidence: Evidence[];
  risks: RiskFlag[];
  unknowns: string[];
  reasoning: string;
  nextAction: string;
}