# lib/api-zod/src/generated/types/performance.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Performance {
  equity: number;
  returnPct: number;
  winRate: number;
  maxDrawdown: number;
  paperTrades: number;
  falsePositiveRate: number;
}