# lib/api-zod/src/generated/types/replayInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReplayInputDataset } from './replayInputDataset';

export interface ReplayInput {
  dataset: ReplayInputDataset;
  /** @minimum 1 */
  startingCapital: number;
  /** @minimum 0 */
  slippageBps: number;
  /** @minimum 0 */
  feeBps: number;
  /** @minimum 0 */
  latencyMs: number;
}