# lib/api-zod/src/generated/types/riskFlag.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RiskFlagSeverity } from './riskFlagSeverity';

export interface RiskFlag {
  severity: RiskFlagSeverity;
  label: string;
  detail: string;
}