# lib/api-zod/src/generated/types/score.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Score {
  opportunity: number;
  risk: number;
  confidence: number;
}