# lib/api-zod/src/generated/types/simulation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationStatus } from './simulationStatus';

export interface Simulation {
  id: string;
  label: string;
  dataset: string;
  status: SimulationStatus;
  returnPct: number;
  maxDrawdown: number;
  winRate: number;
  trades: number;
  capital: number;
  createdAt: string;
}