# lib/api-zod/src/generated/types/sourceHealth.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SourceHealthStatus } from './sourceHealthStatus';

export interface SourceHealth {
  name: string;
  status: SourceHealthStatus;
  latencyMs: number;
  lastChecked: string;
}